Fix client integration bug when adding two relations that access the same index (#88)
## Issue
When adding a second relation with the same index, we fail to create the
index and enter a fail/defer/retry loop. This means the user is never
created, so the new relation is never accessible.

## Solution
This PR updates the error response check in `call()` to verify that
`error_response is None`. This fixes the issue because we were getting
false negatives when evaluating `not error_response` in this function,
since a failed response is falsy.