[DPE-2347] Add plugin support
This commit implements
[DPE-2347](https://warthogs.atlassian.net/browse/DPE-2347) and
represents the first work to add plugin management to the operator. The
plugin can be broken into 3 parts: (1) a class representing its
lifecycle status and configuration; (2) a class representing the configurations
to be applied or removed in opensearch.yml or keystore as dictionaries; (3)
 a manager class, responsible to actually run the operational tasks related to
installing, configuring, upgrading and removing plugins; and (4) changes in
charm-related classes to initialize the plugin manager and to call corresponding
steps, e.g. installation, configuration, etc.

This commit implements an abstract class: OpenSearchPluginManager, that
maps plugin lifecycle status and changes. That can be seen as follows:
* installation
* configuration
* removal

This commit also implements OpenSearchPlugin and OpenSearchPluginConfig
classes, responsible for managing and providing the configuration needed to
implement each step above. Finally, this commit changes the charm main code
to call the plugin manager on every config-changed event.