Plugins will receive more information from the charm:
* Charm config
* OpenSearch version

With these options, it is possible to have an external URL passed as configuration parameter to the charm. The plugin class loads that config, resolves the version and returns a custom install_path.

Original functionality persists: if the charm does not implement new install_path property, then it returns plugin name.