[DPE-3243][DPE-3246] Async backup and new list-backups output
This method implements an async backup method: now, the charm will not
wait anymore for the backup action to be finished. Users can check the status
of backups using the list-backups command.

Restore action closes all the target indices and the action will execute
synchronously with the restore process itself. Indices that are closed will not
be writable during that time. If the restore takes longer and the `juju run`
action times out, then the action continues to run in the background.
It is possible to see it happening with `juju status`. OpenSearch opens the
indices after the restore action has been executed.

This PR also fixes the integration tests in backups: now, we wait for the
backup to asynchronously finish its task before moving on to the next ones.

Closed tickets:
* [DPE-3243](https://warthogs.atlassian.net/browse/DPE-3243): list-backups
now contains only relevant information, in this case the backup-id and
its status; the PR removes the "backup-type" column, as it has no
meaning in OpenSearch
* [DPE-3246](https://warthogs.atlassian.net/browse/DPE-3246): backups will
be async and restores will be synchronous. Users can follow the restore
in the juju status. In any case, the indices that are being restored
will be closed, so the user needs to wait.