[DPE-3578] clear backup status (#186)
Move the status messages in `opensearch_backup` module to comply with
`charm.status.{set,clear}`. All text has been moved to
`constant_charms.py`.

This PR also simplifies the exception structure, where the plugins
should log their own errors if that is not supposed to come back to the
plugin manager and the charm itself. The exception
`OpenSearchPluginClusterNotReadyError` has been abandoned in favor of
the default exception for the same type of error in
`constant_exceptions`.

Tests have been updated accordingly.