[DPE-3573][DPE-3579] Fix #164 - check _cluster/settings instead of opensearch.yml for plugins (#177)
Adds a GET /_cluster/settings, which returns the configuration of the
cluster at the moment, including defaults. This way, plugins can check
if they should add their configs and request a restart during the start
cycle of the charm.

The `_is_enabled` is also extended to use the cluster settings instead
of checking the configuration files directly.

Moved `OpenSearchPluginConfig` to extend `pydantic.BaseModel`. It
ensures that configuration is translated to `Dict[str, str]`, in special
the boolean value, that must have a format of: `true` or `false`.

Also closes [DPE-3579](https://warthogs.atlassian.net/browse/DPE-3579):
move the plugin settings status to use the charm's `status.{set,clear}`
APIs.