Remove plugin status message before cluster ready
This PR proposes to move the cluster ready check that originally happens
within `plugin_manager.run()` to be executed at `_on_config_changed`. If
the cluster is ready, then we proceed to inform the user we will start the
plugin configuration.

It also updates the message to inform we are checking the plugin configs
instead of applying anything.