Re-balance roles deterministically instead of randomly (#209)
(for auto-generated roles)

Algorithm:
- if odd # of nodes: all nodes are cluster manager eligible
- if even # of nodes: node with highest unit number is data node, all other nodes are cluster manager eligible

Replaces old algorithm of `random.choice` one of the nodes that should change roles

This will enable in-place upgrades to be ordered from highest to lowest unit number even if roles are re-balanced (for HA [e.g. because of network cut]) while upgrade is in-progress

That allows upgrade/rollback to be coordinated without leader unit setting information in peer databag, which allows partial rollback even if leader unit is in error state

Context:
https://chat.canonical.com/canonical/pl/zpqoyei54ffr5kwgwkbimizhwo
https://warthogs.atlassian.net/browse/DPE-3878

---

Also, this algorithm ensures that if a unit needs to be restarted for HA while upgrade in-progress, a unit (the highest number unit) on the newer version of the charm/workload will be upgraded. This reduces the likelihood that:
- one of the last units on the older workload version restarts and
- the cluster manager switches to operating with the new workload version (i.e. it stops operating in compatability mode with the old version—see https://discuss.elastic.co/t/rolling-upgrades-master-nodes-voting-config-exclusions/320463/2)

which would mean that units on the old version could not re-connect to the cluster