Handle leader raising uncaught exception while acquiring peer databag lock
If the leader unit raises an uncaught exception in the same Juju event that it acquires the peer databag lock, the changes to the peer databag will be reverted—which could allow another unit to get the lock.

Therefore, we should wait until the next Juju event to use the lock.

Note: for the leader unit, instead of the lock being requested & granted in the same event, the process is now:
- leader unit requests lock & provisionally grants itself lock
- non-leader units get peer-relation-changed and put a unique value in their databag to trigger peer-relation-changed on leader
- leader gets peer-relation-changed (or any other event) and proceeds to use lock and start/stop opensearch

Context: https://matrix.to/#/!xdClnUGkurzjxqiQcN:ubuntu.com/$dRO72RvyFD4EI65_n83hIQPAMFtHmVOQtjG5bMENaMk?via=ubuntu.com&via=matrix.org&via=cutefunny.art