Removing docs folder to recalculate index (#348)
## Issue
The sync_docs workflow is failing due to a known issue with page
nesting. This has been fixed on the discourse side, but the workflow
fails before it is able to re-generate the table of context in
`index.md`

## Solution
Deleted the entire `/docs` folder so the action can re-import them from
scratch.