[DPE-4792] reload certificates via API (#380)
## Issue
Currently the OpenSearch application is restarted when TLS certificates
are renewed. This could be simplified by using the "hot reload" API
OpenSearch provides for renewing the TLS certs.

## Solution
- only keep the `fully-qualified hostname` and the internal IP address
in the `Subject Alternative Names` of the certificate signing request
(and thus the TLS certificates) to mitigate upstream OpenSearch bug
https://github.com/opensearch-project/security/issues/4480
- activate the `plugins.security.ssl_cert_reload_enabled` setting
- add the config `plugins.security.authcz.admin_dn` to all units as this
is required to authenticate against the API
- when TLS conf is set, query the API instead of restarting OpenSearch
- `on_config_change`, initiate a restart directly instead of waiting for
the restart after applying the renewed TLS certificates
- add an integration test that triggers renewing TLS certificates by
running the `set-tls-private-key` action; the test checks the TLS certs
are reloaded directly instead of restarting. The reload via API can also
be seen in the debug logs (e.g.
[here](https://github.com/canonical/opensearch-operator/actions/runs/10182985233/job/28167248618?pr=380#step:30:4972))

## Note
- included bugfix: add missing `refresh_relation_data()` for large
deployments `on_secret_changed`