[DPE-4935] no allocation exclusions on restart (#391)
## Issue
When an Opensearch node is restarted by the operator, currently the
shards assigned to this node are first moved away from the node, and
after restarting moved back. This is a lot of work that, in case of
simple restarts, is not necessary.

To avoid this behavior, no allocation exclusion should be added for the
departing node, when a restart operation is performed.

## Solution
Check for the `restart` flag in `_stop_opensearch()` and pass it through
to `opensearch_exclusions.add_current()`. If set to `True`, only the
voting exclusion should be added for the departing node, but no
allocation exclusion.