[DPE-5283] stabilization for CA rotation and integration test for large deployments (#436)
This PR addresses the stabilization of the CA rotation workflow.

The following issues are fixed:
1) Error with `No cert in -in file '/tmp/tmp1uvvhlyo.cert' matches
private key [xxx] certificate routines:X509_check_private_key:key values
mismatch:../crypto/x509/x509_cmp.c:405:`: The private key was
regenerated on every admin certificate request. When a CA rotation is
restarting the nodes, the certificate that was being loaded was
generated using an old private key and can no longer be validated with
the stored key (see
https://github.com/canonical/opensearch-operator/issues/417).

2) A long delay when restarting after renewing the TLS certificates on
CA rotation (see
https://github.com/canonical/opensearch-operator/issues/417).

3) An issue with the CA truststore file being updated/cleaned too early
(before all nodes have been restarted in rolling fashion), which caused
the respective node not to be able to join the cluster anymore after
restarting.

4) An issue with updating the pem-file too early. This file is used for
API requests by the operator to the node on which it is stored. If this
file is updated too early, the operator can not communicate with the
node anymore. The same if it is updated too late in the workflow.

5) An issue with deferring `CertificateAvailableEvents` on node
initialization which caused the deployment in setups with `manual
tls-operator` to end in endless deferrals and never to start correctly.

---------

Co-authored-by: René Radoi <rene.radoi@canonical.com>