[DPE-5483] Update logging on `requests()`code path (#446)
Currently, each call to `OpenSearchDistribution.requests()` or
`OpenSearchCharm.alt_hosts` may generate a number of `ERROR` entries in
the charm log. However, these are checks or attempts executed within the
each of the routines and not necessarily represent an error of the whole
process.

This PR reduces log to `DEBUG` level instead.