Enable cluster manager only nodes (#424)
## Issue
Currently we always add the `data` role to a node if it is
`cluster-manager`. This is required because otherwise the security index
could not be initialized directly after startup of the first node.

## Solution
This PR provides a solution for enabling "cluster-manager-only" nodes in
large deployments. The workaround for adding the `data` role by default
is removed.

The solution is implemented according this workflow:
- when no data node in the application:
-> no node starts, set status to `blocked`
-> the start hook gets deferred endlessly

- when a data node joins - through large deployment relations:
-> we become aware of it through the fleet_apps object
-> we then start the leader unit of the main-orchestrator
-> which then notifies the large deployment relations
-> the leader unit of the data cluster starts too and initializes the
security index
-> start flow of the rest of the fleet resumes as usual (this waits for
the previously deferred `StartEvent` on the main-orchestrator to be
re-emitted)

For this, the data model of the `PeerClusterApp` is adjusted and the
roles of the application are added, in order to be able to check if
there is any `data` role in the entire cluster fleet (can be queried
with `ClusterTopology.data_role_in_cluster_fleet_apps()`)