[DPE-4575] Manage voting exclusions in more detail (#405)
Currently, our charm is very simplified when it comes to managing voting
exclusions. The main issue is that we have a limited number of possible
votes to exclude at a time, set by
`cluster.max_voting_config_exclusions` (default: `10`). Therefore, we
need to be cautious on cleaning up this list periodically.

This PR does the following changes:
* Adds a method to fetch for the current voting exclusions
* Decides which votes should stay, based on a current list of votings to
be deleted
* Move the `delete_current()` to be executed at `_start_opensearch`
(already there, in `_post_start_init`)
* Add integration test to scale from 3->1->3 and check for the exclusion
rules
* Continuous writes now set non balanced indices with "0-all" replicas