[DPE-5460] restart data only nodes without lock (#460)
## Issue
In Large Deployments with cluster-manager-only nodes and data-only
nodes, a deadlock might appear if the whole data-application has been
powered down and needs to restart. In this case, starting the data node
should be possible without acquiring the lock first, because the lock
index is not available.

## Solution
Start the leader of the data-nodes without acquiring the lock if it was
previously started but is not running anymore.

The PR also adds unit test coverage for the case where a
cluster-manager-only has not initialized the security index and a
failover-orchestrator with data-role joins and wants to start (see
[here](https://github.com/canonical/opensearch-operator/pull/424#discussion_r1768132249)
for context).