Separate alerts when cluster is in yellow state
When a cluster is in yellow state, it's because some shards are
not active. This can be reached under heavy load, however if there
are shards unassigned and the cluster is yellow, that means that
some replicas are not allocated and it might be necessary to add
new nodes in order to host all shards.