Add alert if an OpenSearch scrape fails
If a scrape fails, this might indicate that a unit is not in
a healthy state.

OpenSearch right now does not have a metric saying that one node
is down. E.g. If the systemd service is stopped in one node, the
cluster (N nodes) will drop the faulty node because connectivity
issues and the metrics will show that the cluster now has N-1 nodes
without saying that one node has failed.

With this new alert, at least a notification will appear if one
node stop being responsive.