Relation broken managed at base class:
Given the broken event means that active_relation=False, then
we have to manage the relation going away at the level of the base class.

This way, any call will result on the same processing flow.

We also move the BackupManager *status methods to be static. This way,
even the Base class that does not instantiate a BackupManager object can
use them during relation_broken.