[DPE-6597][DPE-6779] Remove peer-cluster observer and Unify S3/Azure integration (#585)
This PR converges the API-related code in the backup components to a
single class, named "BackupManager". This class will handle the
primitives (backup, restore, list, clean, etc) that result in API calls
to the backup system in OpenSearch.

The OpenSearchBackup class itself is then broken into three:
* OpenSearchMainBackup: for the MAIN cluster - handle all common code
between S3 and azure event handling
* OpenSearch{S3,Azure}Backup: for the MAIN cluster - handle all the S3 /
azure event specifics and inherit from OpenSearchMainBackup

The goal is to unify the Azure and S3 backup components as much as
possible.

This PR also removes unneeded code:
* Peer cluster listener: we do not need anymore and we stick exclusively
with secret changed for the opensearch non-main apps
* `OpenSearchPluginDataProvider`: any more complex data management must
be done by upper classes instead of the plugin, in the case of backup,
the `OpenSearchMainBackup` and its children will be the ones handling
this extra data needed
* The switch code within `OpenSearch{S3,Azure}Plugin` that switches
between main and non-main orchestrator: this is also handled within the
`OpenSearchMainBackup` and its children

Closes #606, #623