[DPE-5490] [DPE-6680] [DPE-6694] [DPE-6763] [DPE-6764] Fix edge case failover app removal (#593)
## Issue

When removing or scaling down an orchestrator in a large deployment or
when removing the relations with an orchestrator, several parts of the
code access properties of NoneType objects. Additionally, the failover
promotes itself in cases where it shouldn't.

## Solution
- guard against accessing properties of NoneTypes
- stop the failover from promoting itself when it shouldn't
- treat scaling an orchestrator down to 0 units as deleting the
orchestrator
- put peer-cluster requirer in a `Waiting` state if the main
orchestrator is removed and a failover is related
- put peer-cluster requirer in a `Blocked` state if both the main and
failover orchestrators are removed

This PR fixes #555 #564 #568 #569

Supersedes #440