Large deployment fixes (#644)
Failover promotion [DPE-6713]:
The logic is currently handled by the failover orchestrator when the
relation with main is lost. It could be the case that only the failover
has lost communication with the main orchestrator and so it should not
promote itself in this case.

NoneType properties [DPE-6680] [DPE-6763] [DPE-6764]:
Application/relation removals lead to properties being `NoneTypes` and
caused exceptions to be raised when accessed.

Role-validation [DPE-7017]:
The operator previously required certain parities for quorum to be
maintained. Opensearch can intelligently handle this as long as the
number of cluster-manager-eligible nodes is at least 3.

Failover promotion:
The fleet apps should keep the failover updated on their connection
status with the main orchestrator and the failover should only be
promoted if the majority of the fleet can't see the main orchestrator.

NoneType properties:
Filter out `NoneTypes` before accessing properties

Role-validation:
Change the validate_roles logic to simply check for at least 3
cm-eligible nodes.

[DPE-6713]:
https://warthogs.atlassian.net/browse/DPE-6713?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[DPE-7017]:
https://warthogs.atlassian.net/browse/DPE-7017?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

[DPE-6680]:
https://warthogs.atlassian.net/browse/DPE-6680?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[DPE-6763]:
https://warthogs.atlassian.net/browse/DPE-6763?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ
[DPE-6764]:
https://warthogs.atlassian.net/browse/DPE-6764?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ