[DPE 7933] - Plugin Manager Rework (#690)
This PR simplifies the plugin manager component and adds support for the
notifications plugin

More specifically this PR:
- Removes logic handling the knn_plugin config option
- Simplifies the OpenSearch keystore manager by removing the base class
and custom errors
- Adds a PluginEvents component which handles SMTP events and can be
extended to add other plugin relation event handlers
- Adds a `plugins` field to `PeerClusterRelData` to transfer plugin
secrets
- Updates the backups component to use the new plugin manager api when
adding or removing keys

The logic to transfer secret key entries to units in sub-clusters
differs from how the backup credentials are transferred (that method
doesn't work corrrectly, see #693). The keys are transferred as follows:
- Each app leader maintains a list of plugin secrets in their app data
bags
- Units in subclusters additionally store a list of secrets they are
tracking in their unit data bags
- The main orchestrator leader creates a secret with the keys, grants it
to subclusters, writes the secret id to the peer-cluster-relation data
- Subcluster leaders read the list transferred from main and compute a
diff with their own list and add or remove from the list accordingly
- Subcluster units compute a diff between the secrets they are tracking
and the secrets the app is tracking and add or remove from their list
(and keystores) accordingly.
- The sequence above handles the cases when a secret is created or
removed. Changes that happen between adding or removing the secret are
handled during the `secret-changed` event.
```mermaid
sequenceDiagram
    Main(leader)->>+Main(leader): create_secret(keys)
    Main(leader)->>+Main(leader): state.add_secret_id(secret_id)
    Main(leader)->>+Main(leader): refresh_rel_data
    Main(leader)-->>-peer-cluster: peer-cluster-relation-changed
    peer-cluster->>+peer-cluster: state.add_secret_id(secret_id)
    peer-cluster-->>+opensearch-peers: opensearch-peers-relation-changed
    opensearch-peers->>+opensearch-peers: get_secret(secret_id)
    opensearch-peers->>+opensearch-peers: update_keystore(keys)
```
## How was this change tested?
- [ x] Manually
- [ ] Unit tests
- [ ] Integration tests