[DPE-8527] - extract certificates key/trust stores openssl logic into helpers (#711)
## Description of issue or feature:
The openssl operations on PKCS12 certificates and trust stores related
operations currently in the TLS module will be reused by the backups
modules.

There was a bug in the previous TLS implementation that lost the whole
cert chain when storing in the trust store with the exception of the
leaf certificate.

## Solution:
- Extract those into helpers
- Fix the bug described above by breaking down CA chains into individual
entries in the trust store

## How was this change tested?
- [ ] Manually
- [x] Unit tests
- [x] Integration tests

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran
to -->
<!--- see how your change affects other areas of the code, etc. -->

## Checklist
- [ ] I have added or updated any relevant documentation.
- [ ] I have cleaned any remaining cloud resources from my accounts.
# Conflicts:
#	lib/charms/opensearch/v0/helper_security.py
#	lib/charms/opensearch/v0/opensearch_tls.py